change tax datatype

+++ server/pg_models/counselor_bill_setting.js
@@ -33,7 +33,7 @@ module.exports = (sequelize, DataTypes) => {
             type: DataTypes.STRING
         },
         tax: {
-            type: DataTypes.INTEGER
+            type: DataTypes.REAL
         }
     }
     ,{
